Parse URL When Ensuring CORS for Direct Uploads
The `request.url` returns the full URL, with path included. This isn't
valid for a CORS header, which needs just the scheme, host, and port if
it's non-standard. Update the `DirectUpload.ensure_cors!` method to
parse out those pieces of the URL and re-assemble it for the CORS
header and ID.

No changelog

Fixes #19