Replace App Template Command With Install Generator in Upgrade Docs
In the upgrade guide for v3.4, we're instructing users to apply an app
template which no longer exists. Instead of using the app template, we
now rely on a generator called `workarea:install` to place the expected
files into your Rails app directory, so update the command in docs to
avoid confusion.