Keep `_id` Suffix In Customized Fields
When adding a customized field to a `Customizations` class that ends in
`_id`, Workarea was previously stripping this suffix from the computed
instance variable name that is converted into snake case from any kind
of input. This causes issues because the data doesn't appear to be
making it into customizations, but is really there under a different
instance variable name. To resolve the issue, Workarea is now using the
`#underscore` String helper prior to calling `#optionize`, which will
cause the value to be properly cased before it's displayed to the end
user.

(#144)