Update `Redis::Rack::Cache` to v2.2.0
This new version requires `Rack::Cache` v1.10 and enables over-the-wire
gzip compression to the Redis server. This feature is useful for
extremely high traffic sites, but should be used with caution since it
will increase the CPU/RAM load on your application server. You should
use this if the trade-off between RAM increase and bandwidth decrease
makes sense.

WORKAREA-94